pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "milvus-admin-ui"
version = "0.3.0"
description = "Milvus admin & RAG UI with S3-compatible data staging utilities"
requires-python = ">=3.11,<3.12"
readme = "README.md"
authors = [{ name = "milvus-admin-ui", email = "team@example.com" }]
license = { text = "Apache-2.0" }
keywords = ["ibm-cos", "s3", "milvus", "vector", "rag", "admin-ui"]

dependencies = [
  "ibm-cos-sdk>=2.13.0",
  "boto3>=1.34.0",
  "python-dotenv>=1.0.1",
  "tenacity>=8.2.3",
  "pymilvus>=2.4.6",
  "numpy>=1.26",
  "sentence-transformers>=3.0",
  "fastapi>=0.110",
  "uvicorn>=0.29",
  "pydantic>=2,<3",

  # ✅ Required for FastAPI form-data / file uploads (/api/ingest/upload)
  "python-multipart>=0.0.9",

  # ✅ Recommended for efficient static file and upload streaming
  "aiofiles>=23.2.1"
]

[project.scripts]
mui-ingest = "utils.cli_ingest:main"
mui-create-vectordb = "utils.create_vector_db:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
